refactor: move to TestDebugDescriptions
